requirements.txt
asgiref==3.8.1
boto3==1.34.113
botocore==1.34.113
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.9.0
django-oauth-toolkit==2.4.0
django-ses==4.1.0
django-storages==1.14.3
djangorestframework==3.15.1
idna==3.7
jmespath==1.0.1
jwcrypto==1.5.6
oauthlib==3.2.2
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.1
pycparser==2.22
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.2
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
